+++ kernel/arch/x86/irq.c
@@ -154,12 +154,18 @@ static void _init_timer_source(bool apic_ok) {
 
     if (apic_ok && apic_timer_init(timer_hz)) {
         use_apic_timer = true;
-        log_info("APIC timer initialized at %u Hz", apic_timer_hz());
+        log_info(
+            "APIC timer initialized at %u Hz",
+            (unsigned int)apic_timer_hz()
+        );
         return;
     }
 
     pit_set_frequency(timer_hz);
-    log_info("PIT timer initialized at %u Hz", pit_get_frequency());
+    log_info(
+        "PIT timer initialized at %u Hz",
+        (unsigned int)pit_get_frequency()
+    );
 }
 
 bool irq_init(void) {
@@ -222,7 +228,10 @@ void irq_register(size_t irq, int_handler_t handler) {
         if (irq != IRQ_SYSTEM_TIMER || !use_apic_timer) {
             u32 dest = lapic_id();
             if (!ioapic_route_irq((u8)irq, (u8)vec, dest)) {
-                log_warn("failed to route irq %u via IOAPIC", (u32)irq);
+                log_warn(
+                    "failed to route irq %u via IOAPIC",
+                    (unsigned int)irq
+                );
             }
         }
         return;
